+++ wazimap_ng/profile/serializers/profile_indicator_serializer.py
@@ -5,7 +5,7 @@ from rest_framework import serializers
 from wazimap_ng.datasets.models import IndicatorData
 from wazimap_ng.profile.models import ProfileIndicator
 
-from .indicator_data_serializer import get_indicator_data
+from .utils import get_indicator_data
 
 import logging
